chore: :label: change int parameters in bubble_options to be double
It is more convenient to have these parameters as `double`, but the dependant library uses `int` for these values, however when casting these values from `double` to `int`, there will not be any sensable change, so these values sent as `double`s from the dart side and gets converted them to `int`s in the native side

+++ android/src/main/kotlin/dev/moaz/dash_bubble/src/BubbleService.kt
@@ -1,4 +1,3 @@
-
 package dev.moaz.dash_bubble.src
 
 import android.app.Notification
@@ -67,11 +66,22 @@ class BubbleService : FloatingBubbleService() {
         )
 
         return FloatingBubble.Builder(this)
-            .bubble(bubbleIcon, bubbleOptions.bubbleSize!!, bubbleOptions.bubbleSize!!)
+            .bubble(
+                bubbleIcon,
+                bubbleOptions.bubbleSize!!.toInt(),
+                bubbleOptions.bubbleSize!!.toInt()
+            )
             .bubbleStyle(null)
-            .startLocation(bubbleOptions.startLocationX!!, bubbleOptions.startLocationY!!)
+            .startLocation(
+                bubbleOptions.startLocationX!!.toInt(),
+                bubbleOptions.startLocationY!!.toInt()
+            )
             .enableAnimateToEdge(bubbleOptions.enableAnimateToEdge!!)
-            .closeBubble(closeIcon, bubbleOptions.bubbleSize!!, bubbleOptions.bubbleSize!!)
+            .closeBubble(
+                closeIcon,
+                bubbleOptions.bubbleSize!!.toInt(),
+                bubbleOptions.bubbleSize!!.toInt()
+            )
             .closeBubbleStyle(null)
             .enableCloseBubble(bubbleOptions.enableClose!!)
             .bottomBackground(bubbleOptions.enableBottomShadow!!)
